# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Made Super AI Agent API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  /** @nullable */
  lastMessage?: string | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
}

export interface ConversationUpdate {
  title?: string;
  lastMessage?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface MessagePair {
  userMessage: Message;
  assistantMessage: Message;
}

export interface StatsSummary {
  totalConversations: number;
  totalMessages: number;
  todayMessages: number;
  recentConversations?: Conversation[];
}

export type SocialAccountPlatform = typeof SocialAccountPlatform[keyof typeof SocialAccountPlatform];


export const SocialAccountPlatform = {
  twitter: 'twitter',
  youtube: 'youtube',
  instagram: 'instagram',
  tiktok: 'tiktok',
  facebook: 'facebook',
  linkedin: 'linkedin',
} as const;

export interface SocialAccount {
  id: number;
  platform: SocialAccountPlatform;
  username: string;
  connected: boolean;
  /** @nullable */
  apiKey?: string | null;
  createdAt: string;
}

export type SocialAccountInputPlatform = typeof SocialAccountInputPlatform[keyof typeof SocialAccountInputPlatform];


export const SocialAccountInputPlatform = {
  twitter: 'twitter',
  youtube: 'youtube',
  instagram: 'instagram',
  tiktok: 'tiktok',
  facebook: 'facebook',
  linkedin: 'linkedin',
} as const;

export interface SocialAccountInput {
  platform: SocialAccountInputPlatform;
  username: string;
  apiKey?: string;
  apiSecret?: string;
  accessToken?: string;
}

export type SocialPostStatus = typeof SocialPostStatus[keyof typeof SocialPostStatus];


export const SocialPostStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
  failed: 'failed',
} as const;

export type SocialPostPostType = typeof SocialPostPostType[keyof typeof SocialPostPostType];


export const SocialPostPostType = {
  text: 'text',
  video: 'video',
  image: 'image',
} as const;

export interface SocialPost {
  id: number;
  platform: string;
  content: string;
  status: SocialPostStatus;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  postType?: SocialPostPostType;
  /** @nullable */
  videoScript?: string | null;
  createdAt: string;
}

export type SocialPostInputStatus = typeof SocialPostInputStatus[keyof typeof SocialPostInputStatus];


export const SocialPostInputStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
} as const;

export type SocialPostInputPostType = typeof SocialPostInputPostType[keyof typeof SocialPostInputPostType];


export const SocialPostInputPostType = {
  text: 'text',
  video: 'video',
  image: 'image',
} as const;

export interface SocialPostInput {
  platform: string;
  content: string;
  status?: SocialPostInputStatus;
  scheduledAt?: string;
  postType?: SocialPostInputPostType;
  videoScript?: string;
}

export type SocialPostUpdateStatus = typeof SocialPostUpdateStatus[keyof typeof SocialPostUpdateStatus];


export const SocialPostUpdateStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
  failed: 'failed',
} as const;

export interface SocialPostUpdate {
  content?: string;
  status?: SocialPostUpdateStatus;
  scheduledAt?: string;
}

export type ContentGenerateInputContentType = typeof ContentGenerateInputContentType[keyof typeof ContentGenerateInputContentType];


export const ContentGenerateInputContentType = {
  post: 'post',
  video_script: 'video_script',
  caption: 'caption',
  thread: 'thread',
} as const;

export type ContentGenerateInputTone = typeof ContentGenerateInputTone[keyof typeof ContentGenerateInputTone];


export const ContentGenerateInputTone = {
  professional: 'professional',
  casual: 'casual',
  energetic: 'energetic',
  educational: 'educational',
} as const;

export interface ContentGenerateInput {
  topic: string;
  platform: string;
  contentType?: ContentGenerateInputContentType;
  tone?: ContentGenerateInputTone;
}

export interface GeneratedContent {
  content: string;
  platform: string;
  contentType: string;
  suggestions?: string[];
}

export type SocialStatsPlatformBreakdownItem = {
  platform: string;
  count: number;
};

export interface SocialStats {
  totalPosts: number;
  scheduled: number;
  published: number;
  connectedAccounts: number;
  platformBreakdown?: SocialStatsPlatformBreakdownItem[];
}

export interface EmailStatus {
  connected: boolean;
  /** @nullable */
  fromAddress?: string | null;
}

export type EmailGenerateInputTone = typeof EmailGenerateInputTone[keyof typeof EmailGenerateInputTone];


export const EmailGenerateInputTone = {
  professional: 'professional',
  casual: 'casual',
  friendly: 'friendly',
  formal: 'formal',
} as const;

export interface EmailGenerateInput {
  /** @minLength 1 */
  prompt: string;
  tone?: EmailGenerateInputTone;
}

export interface EmailDraft {
  subject: string;
  body: string;
}

export interface EmailSendInput {
  /** @minLength 3 */
  to: string;
  subject: string;
  /** @minLength 1 */
  body: string;
}

export interface SentEmail {
  id: number;
  toAddress: string;
  subject: string;
  body: string;
  /** @nullable */
  fromAddress?: string | null;
  createdAt: string;
}