# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Made Super AI Agent API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /email/generate:
    post:
      operationId: generateEmail
      tags: [email]
      summary: AI-draft an email (subject + body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailGenerateInput"
      responses:
        "200":
          description: Drafted email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
  /email/send:
    post:
      operationId: sendEmail
      tags: [email]
      summary: Send an email from the connected account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailSendInput"
      responses:
        "201":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SentEmail"
...
components:
  schemas:
...
    EmailGenerateInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
        tone:
          type: string
          enum: [professional, casual, friendly, formal]
    EmailDraft:
      type: object
      required: [subject, body]
      properties:
        subject:
          type: string
        body:
          type: string
    EmailSendInput:
      type: object
      required: [to, subject, body]
      properties:
        to:
          type: string
          minLength: 3
        subject:
          type: string
        body:
          type: string
          minLength: 1
    SentEmail:
      type: object
      required: [id, toAddress, subject, body, createdAt]
      properties:
        id:
          type: integer
        toAddress:
          type: string
        subject:
          type: string
        body:
          type: string
        fromAddress:
          type: ["string", "null"]
        createdAt:
          type: string