# lib/api-zod/src/generated/types/emailGenerateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Made Super AI Agent API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailGenerateInputTone } from './emailGenerateInputTone';

export interface EmailGenerateInput {
  /** @minLength 1 */
  prompt: string;
  tone?: EmailGenerateInputTone;
}