# lib/api-zod/src/generated/types/emailSendInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Made Super AI Agent API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailSendInput {
  /** @minLength 3 */
  to: string;
  subject: string;
  /** @minLength 1 */
  body: string;
}