# lib/api-zod/src/generated/types/sentEmail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Made Super AI Agent API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SentEmail {
  id: number;
  toAddress: string;
  subject: string;
  body: string;
  /** @nullable */
  fromAddress?: string | null;
  createdAt: string;
}